pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "Elias Chokeir", email = "ltpie12345+github@gmail.com" }]
classifiers = [
   "Development Status :: 4 - Beta",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
   "fastapi[all]>=0.115.8",
   "pydantic-settings>=2.8.0",
   "python-dotenv>=1.0.1",
   "sqlalchemy>=2.0.38",
   "sqlmodel>=0.0.22",
]

[project.urls]
Documentation = "https://github.com/Elias Chokeir/app#readme"
Issues = "https://github.com/Elias Chokeir/app/issues"
Source = "https://github.com/Elias Chokeir/app"

[tool.hatch.version]
path = "src/app/__about__.py"

[tool.hatch.envs.default]
# Default environment
python = "3.12"
installer = "uv"
path = ".venv"

[tool.hatch.envs.default.scripts]
# Default scripts for dev and database operations
dev = "fastapi dev {args:src/app/main.py}"
prod = "fastapi run"

[tool.hatch.envs.types]
installer = "uv"
extra-dependencies = ["mypy>=1.0.0"]

[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/app tests}"

[tool.coverage.run]
source_pkgs = ["app", "tests"]
branch = true
parallel = true
omit = ["src/app/__about__.py"]

[tool.coverage.paths]
app = ["src/app", "*/app/src/app"]
tests = ["tests", "*/app/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
